extremely huge change (atleast for me).
What's new:
added a char filter for every single program. everytime a program asks for a number or integer, and a user inputs special characters, the program ignores all but integers so to prevent abrupt errors from terminating the program.

the filter is very barebones and done on the spot, but it took hours to make so im really proud of that one.

aside from that, minor adjustments here and there.

What's next:
just need to finish my list of descriptions for every program and possibly polish a little(hopefully).